RUB-243: extract utxo_u128, mempool_stats, miner_coinbase — block extraction
Pure block extraction, zero logic changes.

- utxo_basic.go: -39 lines (-11 complexity) → utxo_u128.go: +40 lines (+10 complexity)
- mempool.go: -256 lines (-26 complexity) → mempool_stats.go: +263 lines (+27 complexity)
- miner.go: -118 lines (-26 complexity) → miner_coinbase.go: +126 lines (+27 complexity)

Net complexity: +1 (within ≤20 gate)

parity-exempt: code organization only, no Rust counterpart.
